Fix existing Windows build by moving bundled GLEW files into "GL" subdir.
