+++ msvc/GL/glew.c
@@ -31,7 +31,7 @@
 */
 
 #ifndef GLEW_INCLUDE
-#include "glew.h"
+#include "GL/glew.h"
 #else
 #include GLEW_INCLUDE
 #endif
@@ -50,7 +50,7 @@
 #  if defined(NOGDI)
 #    undef NOGDI
 #  endif
-#  include "wglew.h"
+#  include "GL/wglew.h"
 #elif !defined(__ANDROID__) && !defined(__native_client__) && !defined(__HAIKU__) && (!defined(__APPLE__) || defined(GLEW_APPLE_GLX))
 #  include <GL/glxew.h>
 #endif
