+++ msvc/GL/glewinfo.c
