+++ msvc/projectM.vcxproj
@@ -133,8 +133,8 @@
     <ClCompile Include="$(MSBuildThisFileDirectory)../src\libprojectM\timer.cpp" />
     <ClCompile Include="$(MSBuildThisFileDirectory)../src\libprojectM\wipemalloc.cpp" />
     <ClCompile Include="$(MSBuildThisFileDirectory)../src\libprojectM\FileScanner.cpp" />
-    <ClInclude Include="$(MSBuildThisFileDirectory)../msvc\glew.h" />
-    <ClCompile Include="$(MSBuildThisFileDirectory)../msvc\glew.c">
+    <ClInclude Include="$(MSBuildThisFileDirectory)../msvc\GL\glew.h" />
+    <ClCompile Include="$(MSBuildThisFileDirectory)../msvc\GL\glew.c">
       <CompileAs Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">CompileAsC</CompileAs>
       <CompileAs Condition="'$(Configuration)|$(Platform)'=='Release|x64'">CompileAsC</CompileAs>
     </ClCompile>
